requirements.txt
numpy==1.23.0
pandas==1.4.3
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
